fix: use UnixFS for traversal on the Server
The traversal of the DAG depends on which file request was made (e.g.
file contents vs. directory listing). Modify js-ipld so that it
pushes all the nodes visited to the Consumer. This way no custom
code is needed and the Client has all the nodes that are needed
(even for verification).